refactor(#335): review API 함수 폴더 위치 변경
