+++ src/lib/apis/instructorPostApis.ts
@@ -1,8 +1,4 @@
-import {
-  instructorPostResponse,
-  IInstructorReviewList,
-  InstructorUpdate,
-} from '@/types/instructor';
+import { instructorPostResponse, InstructorUpdate } from '@/types/instructor';
 import { FetchError } from '@/types/types';
 
 export const getInstructorPost = async (
@@ -16,34 +12,6 @@ export const getInstructorPost = async (
     return response.data.lecturerProfile;
   } catch (error) {
     return new Error('잘못된 요청입니다!');
-  }
-};
-
-export const getReviews = async (
-  id: string,
-  displayCount: number,
-  currentPage: number,
-  targetPage: number,
-  firstItemId: number,
-  lastItemId: number,
-  lecturerReviewType: string,
-): Promise<{ review: IInstructorReviewList[] } | Error> => {
-  const query = `displayCount=${displayCount}&currentPage=${currentPage}&targetPage=${targetPage}&firstItemId=${firstItemId}&lastItemId=${lastItemId}&lecturerReviewType=${lecturerReviewType}`;
-
-  try {
-    const response = await fetch(
-      `/api/post/review/instructor?id=${id}&${query}`,
-      {
-        credentials: 'include',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      },
-    ).then((data) => data.json());
-
-    return response.data;
-  } catch (error) {
-    return new Error('강사 리뷰 조회 요청 오류!');
   }
 };
